# docs_capsule_zero/adr/openapi.yaml
openapi: 3.1.0
info:
  title: Capsule Zero API
  version: 0.1.0
  description: >
    Sprint 0 implementation contract for the v0.1 REST boundary. This file is
    the source for generated TypeScript clients consumed by both the Next.js web frontend and the React Native mobile app.
...
servers:
  - url: https://{host}
    variables:
      host:
        default: app.capsule-zero.local
security:
  - bearerSession: []
  - sessionToken: []
...
components:
  securitySchemes:
    bearerSession:
      type: http
      scheme: bearer
      bearerFormat: Kratos session token
      description: >
        Session token returned by `/api/auth/login` or `/api/auth/registration`
        and forwarded as `Authorization: Bearer <token>`.
    sessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
      description: >
        Alternate server-to-server session token header consumed by the Go API.
    lavaWebhookKey:
      type: apiKey
      in: header
      name: X-Api-Key
...
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Kratos session is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource was not found or is not visible to the user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequests:
      description: The per-client request budget is exhausted temporarily.
      headers:
        Retry-After:
          description: Seconds until the client should retry.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Temporary server-side failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: Upstream identity service is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: The feature is disabled or a required runtime dependency is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Locale:
      type: string
      enum: [en, ru]
...
    AuthResponse:
      type: object
      required: [requiresEmailConfirmation]
      properties:
        session:
          $ref: "#/components/schemas/AuthSession"
        user:
          $ref: "#/components/schemas/AuthUser"
        profile:
          $ref: "#/components/schemas/Profile"
        requiresEmailConfirmation:
          type: boolean
        verificationFlowId:
          type: string
          description: >-
            Present on registration when Kratos started an email verification
            flow for the new address; the emailed code must be submitted
            against this flow via /api/auth/verification/complete.
    AuthRegistrationRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
          maxLength: 80
        locale:
          $ref: "#/components/schemas/Locale"
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthProvidersResponse:
      type: object
      required: [google]
      properties:
        google:
          type: boolean
          description: Whether this deployment offers Google sign-in (spec 037).
    AuthGoogleStartRequest:
      type: object
      required: [returnTo]
      properties:
        returnTo:
          type: string
          format: uri
          description: >-
            App callback URL Kratos redirects the browser to after the
            provider dance; must be inside the Kratos allowed_return_urls.
    AuthGoogleStartResponse:
      type: object
      required: [redirectUrl, exchangeCode]
      properties:
        redirectUrl:
          type: string
          format: uri
          description: Google consent-screen URL the browser must visit.
        exchangeCode:
          type: string
          description: >-
            Session-token exchange (init) code; the completion call presents
            it together with the callback's return_to code.
    AuthGoogleCompleteRequest:
      type: object
      required: [exchangeCode, returnToCode]
      properties:
        exchangeCode:
          type: string
        returnToCode:
          type: string
          description: Code Kratos appended to the callback redirect.
    AuthRecoveryRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    AuthRecoveryResponse:
      type: object
      required: [delivery, email, flowId]
      properties:
        delivery:
          type: string
          enum: [email]
        email:
          type: string
          format: email
        flowId:
          type: string
          description: >-
            Kratos flow the emailed one-time code is bound to; completion must
            submit the code against this flow.
    AuthRecoveryCompleteRequest:
      type: object
      required: [newPassword]
      properties:
        flowId:
          type: string
          description: Required with `code` for the first completion attempt.
        code:
          type: string
          description: Required with `flowId` for the first completion attempt.
        newPassword:
          type: string
          minLength: 8
        recoveryContinuationId:
          type: string
          description: >-
            Opaque retry id returned in `error.details.recoveryContinuationId`
            when the recovery code was valid but the submitted password was
            rejected by Kratos policy. Retrying with this id lets the user
            choose a different password without consuming the already-used
            recovery code again.
    AuthVerificationRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    AuthVerificationResponse:
      type: object
      required: [delivery, email, flowId]
      properties:
        delivery:
          type: string
          enum: [email]
        email:
          type: string
          format: email
        flowId:
          type: string
    AuthVerificationCompleteRequest:
      type: object
      required: [flowId, code]
      properties:
        flowId:
          type: string
        code:
          type: string
    AuthPasswordChangeRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
    AuthOkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    AuthLogoutResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    Profile:
      type: object
      required:
        [userId, email, displayName, locale, coinBalance, createdAt, updatedAt]
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        avatarUrl:
          type: string
        locale:
          $ref: "#/components/schemas/Locale"
        country:
          type: string
        city:
          type: string
        coinBalance:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProfileUpdateRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        locale:
          $ref: "#/components/schemas/Locale"
        # Set-or-leave semantics: an omitted field keeps the stored value. `null`
        # is intentionally not accepted — the Go handler treats a missing pointer
        # as "leave unchanged" (COALESCE), so advertising a nullable "clear" that
        # the API silently ignores would mislead clients. Clearing a saved field
        # is a deferred feature (see spec 024 Known Issues).
        country:
          type: string
          minLength: 1
          maxLength: 80
        city:
          type: string
          minLength: 1
          maxLength: 80
        avatarUrl:
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
          description: Absolute https URL of the avatar image.
...
    PhotoUploadInitRequest:
      type: object
      required: [fileName, contentType, sizeBytes]
      properties:
        fileName:
          type: string
          minLength: 1
          maxLength: 255
          pattern: "^(?!\\.\\.?$)[^/\\\\]+$"
          description: Original basename for validation/audit only; never used as the object key.
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        sizeBytes:
          type: integer
          minimum: 1
          maximum: 10485760
    PhotoUploadInitResponse:
      type: object
      description: >-
        Direct-upload contract. Clients must send every returned upload header.
        The presigned URL is sensitive and may reveal provider routing details,
        the bucket name, and the server-generated object key; never log it.
      required: [jobId, assetId, uploadUrl, uploadHeaders, expiresAt]
      properties:
        jobId:
          type: string
          format: uuid
        assetId:
          type: string
          format: uuid
        uploadUrl:
          type: string
          format: uri
        uploadHeaders:
          type: object
          additionalProperties:
            type: string
          description: Exact request headers signed into the PUT URL, including Content-Type.
        expiresAt:
          type: string
          format: date-time
    PhotoUploadCompleteRequest:
      type: object
      required: [jobId, assetId]
      properties:
        jobId:
          type: string
          format: uuid
        assetId:
          type: string
          format: uuid
...
paths:
  /api/auth/registration:
    post:
      tags: [Auth]
      operationId: registerWithPassword
      x-auth: public
      security: []
      summary: Create a password account and return a session or email-confirmation state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegistrationRequest"
      responses:
        "200":
          description: Registration completed, or email confirmation is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
        "500":
          $ref: "#/components/responses/InternalServerError"
        default:
          $ref: "#/components/responses/InternalServerError"
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: loginWithPassword
      x-auth: public
      security: []
      summary: Establish a password session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Authenticated session and mapped profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/auth/whoami:
    get:
      tags: [Auth]
      operationId: getCurrentSession
      x-auth: public
      security: []
      summary: Resolve the presented session token, or return an empty auth response.
      responses:
        "200":
          description: Current session response; empty when no valid session is presented.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
        default:
          $ref: "#/components/responses/InternalServerError"
  /api/auth/recovery:
    post:
      tags: [Auth]
      operationId: requestPasswordRecovery
      x-auth: public
      security: []
      summary: Start code-method password recovery without revealing whether the account exists.
      description: >-
        Sends a one-time recovery code by email (Kratos code method, spec 035)
        and returns the flow the code is bound to. The response shape is
        identical for known and unknown addresses.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRecoveryRequest"
      responses:
        "200":
          description: Recovery request accepted for email delivery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthRecoveryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/auth/recovery/complete:
    post:
      tags: [Auth]
      operationId: completePasswordRecovery
      x-auth: public
      security: []
      summary: Exchange the emailed recovery code for a new password and a session.
      description: >-
        Submits the one-time code against its flow, sets the new password, and
        signs the user in (spec 035). A wrong or expired code is a validation
        error and never issues a session. If Kratos accepts the code but rejects
        the password policy, the `400 VALIDATION_ERROR` response includes
        `error.details.recoveryContinuationId`; clients retry with that id and
        a different `newPassword` instead of resubmitting the consumed code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRecoveryCompleteRequest"
      responses:
        "200":
          description: Password replaced; authenticated session returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/auth/verification:
    post:
      tags: [Auth]
      operationId: startEmailVerification
      x-auth: public
      security: []
      summary: Start or resend a code-method email verification without revealing whether the account exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerificationRequest"
      responses:
        "200":
          description: Verification request accepted for email delivery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthVerificationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/auth/verification/complete:
    post:
      tags: [Auth]
      operationId: completeEmailVerification
      x-auth: public
      security: []
      summary: Confirm the email address with the emailed verification code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerificationCompleteRequest"
      responses:
        "200":
          description: Address verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthOkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/auth/password:
    post:
      tags: [Auth]
      operationId: changePassword
      x-auth: session
      summary: Change the password of the signed-in user (current password required).
      description: >-
        Verifies the current password by re-authenticating (which also mints
        the privileged session Kratos requires) before setting the new one
        via the settings flow (spec 035).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthPasswordChangeRequest"
      responses:
        "200":
          description: Password changed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthOkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/auth/providers:
    get:
      tags: [Auth]
      operationId: getAuthProviders
      x-auth: public
      security: []
      summary: Report which social sign-in providers this deployment offers.
      responses:
        "200":
          description: Provider availability map.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthProvidersResponse"
        default:
          $ref: "#/components/responses/InternalServerError"
  /api/auth/google/start:
    post:
      tags: [Auth]
      operationId: startGoogleSignIn
      x-auth: public
      security: []
      summary: Begin the native Google OIDC flow (spec 037).
      description: >-
        Creates a Kratos API login flow with a session-token exchange code and
        returns the Google consent URL for the browser plus the exchange code
        the completion call must present. 404 while Google sign-in is
        disabled for the deployment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthGoogleStartRequest"
      responses:
        "200":
          description: Consent redirect URL and exchange code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthGoogleStartResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/auth/google/complete:
    post:
      tags: [Auth]
      operationId: completeGoogleSignIn
      x-auth: public
      security: []
      summary: Exchange the Google callback codes for an authenticated session.
      description: >-
        Trades the start call's exchange code plus the return_to code from the
        Kratos callback redirect for a session token; answers with the
        standard auth response. Wrong, expired, or reused codes are rejected
        with GOOGLE_SIGN_IN_FAILED and never issue a session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthGoogleCompleteRequest"
      responses:
        "200":
          description: Authenticated session and mapped profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      x-auth: public
      security: []
      summary: Revoke the presented session token; idempotent when absent.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "200":
          description: Logout request completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLogoutResponse"
        "502":
          $ref: "#/components/responses/BadGateway"
        default:
          $ref: "#/components/responses/BadGateway"
...
  /api/profile:
    get:
      tags: [Profile]
      operationId: getProfile
      x-auth: user
      summary: Read current profile.
      responses:
        "200":
          description: Current profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [Profile]
      operationId: updateProfile
      x-auth: user
      summary: Update current profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/uploads/photo/init:
    post:
      tags: [Uploads]
      operationId: initPhotoUpload
      x-auth: user
      summary: Validate metadata and return a presigned upload target.
      description: >-
        The default-off feature gate runs before session resolution and returns
        503 FEATURE_UNAVAILABLE while uploads are disabled. When enabled,
        signing/readiness can return 503 STORAGE_UNAVAILABLE and session
        resolution can return 502 when the identity service is unavailable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoUploadInitRequest"
      responses:
        "200":
          description: Upload target.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUploadInitResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/uploads/photo/complete:
    post:
      tags: [Uploads]
      operationId: completePhotoUpload
      x-auth: user
      summary: Verify and complete an initialized photo upload.
      description: >-
        The default-off feature gate runs before session resolution and returns
        503 FEATURE_UNAVAILABLE while uploads are disabled. When enabled,
        object verification can return 503 STORAGE_UNAVAILABLE and session
        resolution can return 502 when the identity service is unavailable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoUploadCompleteRequest"
      responses:
        "200":
          description: Upload job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUploadCompleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...